Remove old grants index page, flatten routers
