+++ middleware/locals.js
@@ -49,7 +49,6 @@ module.exports = {
          */
         res.locals.enablePrompt = config.get('features.enablePrompt');
         res.locals.enableSurvey = config.get('features.enableSurvey');
-        res.locals.enableNewPastGrantsSearch = config.get('features.enableNewPastGrantsSearch');
 
         /**
          * Metadata (e.g. global title, description)
