+++ views/components/hero.njk
@@ -22,11 +22,11 @@
             </div>
             {% if image.caption %}
                 <span class="hero__caption u-desktop-only">
-                    {% if image.grantId and enableNewPastGrantsSearch %}
+                    {% if image.grantId %}
                         <a href="{{ localify("/funding/grants/" + image.grantId) }}" class="u-link-unstyled">
                     {% endif %}
                     <span class="u-caption">{{ image.caption }}</span>
-                    {% if image.grantId and enableNewPastGrantsSearch %}</a>{% endif %}
+                    {% if image.grantId %}</a>{% endif %}
                 </span>
             {% endif %}
         </div>
